Update README.md
Train Rearrangement

There is a T-shaped railroad. Imagine that trains coming from (a), passing by (b), going to (c). We want to re-order the trains. For example, The order of coming trains is 3214. Rearranging the sequence of trains to 1234 is possible by following sequence of operations.

1.

No. 3 train to (b).

2.

No. 2 train to (b).

3.

No. 1 train to (b).